@@ -1,5 +1,5 @@
-const pluginutils = require('@rollup/pluginutils');
-const path = require('path');
+import pluginutils from '@rollup/pluginutils';
+import path from 'path';
 
 const defaults = {
 	size: 150,
