@@ -1,12 +1,12 @@
-const fs = require('fs');
-const os = require('os');
-const path = require('path');
-const mkdirp = require('mkdirp');
-const puppeteer = require('puppeteer');
+import { writeFileSync } from 'fs';
+import { tmpdir } from 'os';
+import { join } from 'path';
+import mkdirp from 'mkdirp';
+import puppeteer from 'puppeteer';
 
-const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
+const DIR = join(tmpdir(), 'jest_puppeteer_global_setup');
 
-module.exports = async function () {
+export default async function () {
 	const browser = await puppeteer.launch();
 	// store the browser instance so we can teardown it later
 	// this global is only available in the teardown but not in TestEnvironments
@@ -14,5 +14,5 @@ module.exports = async function () {
 
 	// use the file system to expose the wsEndpoint for TestEnvironments
 	mkdirp.sync(DIR);
-	fs.writeFileSync(path.join(DIR, 'wsEndpoint'), browser.wsEndpoint());
-};
+	writeFileSync(join(DIR, 'wsEndpoint'), browser.wsEndpoint());
+}
