@@ -20,13 +20,18 @@ import browserslist from 'browserslist';
 import { getFlagEnv, hashf, unifyPath } from './utils';
 import chalk from 'chalk';
 import rimraf from 'rimraf';
-import { glob } from 'glob';
+import glob from 'glob';
 import { preactDebug } from './preact-debug-plugin';
 import typescript from 'rollup-plugin-typescript2';
 import iife from 'rollup-plugin-iife';
 import { minify as terserMinify } from 'terser';
-
+import { fileURLToPath } from 'url';
 export * from './puppeteer';
+
+const { babel } = babelPlugin;
+
+// __dirname is not defined in ES module scope
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const browsers = browserslist();
 
@@ -589,7 +594,7 @@ function getPluginsConfig(defaults, override = []) {
 	const fns = {
 		alias,
 		'node-resolve': nodeResolve,
-		babel: babelPlugin,
+		babel,
 		styles,
 		replace,
 		commonjs,
