@@ -1,6 +1,8 @@
-import { yellow, green } from 'chalk';
+import chalk from 'chalk';
 import puppeteerCore from 'puppeteer-core';
 import { getFlagEnv } from './utils';
+
+const { yellow, green } = chalk;
 
 const isTest = getFlagEnv('TEST_ENV');
 let initial = true;
