@@ -1,7 +1,10 @@
+import { bundler, openPage } from '../../lib/bundler';
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
 jest.setTimeout(20000);
-const { bundler, openPage } = require('../../dist/lib/bundler');
-const fs = require('fs');
-const path = require('path');
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const config = { ...global.configDefaults, testPath: __dirname };
 
 describe('Configuration options', () => {
